+++ src/services/auth/auth-service.ts
@@ -1,0 +1,210 @@
+import { apiConfig } from '@/config';
+import {
+  GoogleAuthResponse,
+  RegisterResponse,
+  LoginResponse,
+  CheckAuthResponse,
+} from '@/types/api';
+import { updateAuthToken } from '../utils/http';
+
+/**
+ * 獲取 Google 登入 URL
+ */
+export const fetchGoogleAuthUrl = async (): Promise<string> => {
+  try {
+    console.log(`發送請求: GET ${apiConfig.baseUrl}/auth/google/auth`);
+    const res = await fetch(`${apiConfig.baseUrl}/auth/google/auth`);
+    console.log('回應狀態:', res.status, res.statusText);
+
+    if (!res.ok) {
+      throw new Error(`HTTP error! status: ${res.status}`);
+    }
+
+    const data: GoogleAuthResponse = await res.json();
+    console.log('回應資料:', data);
+    return data.redirectUri;
+  } catch (error) {
+    console.error('獲取 Google 登入 URL 失敗:', error);
+    throw error;
+  }
+};
+
+/**
+ * 使用 Google 授權碼換取 token
+ */
+export const exchangeGoogleCodeForToken = async (
+  code: string,
+): Promise<any> => {
+  try {
+    // 直接調用後端 API，而不是 Next.js API 路由
+    const res = await fetch(
+      `${apiConfig.baseUrl}/auth/google/callback?code=${encodeURIComponent(code)}`,
+      {
+        method: 'GET',
+      },
+    );
+
+    console.log('回應狀態:', res.status, res.statusText);
+
+    if (!res.ok) {
+      throw new Error(`HTTP error! status: ${res.status}`);
+    }
+
+    const data = await res.json();
+    console.log('回應資料:', data);
+    return data;
+  } catch (error) {
+    console.error('Google code 換取 token 失敗:', error);
+    throw error;
+  }
+};
+
+/**
+ * 檢查使用者認證狀態並取得新的 Token 與使用者資料
+ * @returns 包含新 Token 和使用者資料的回應，若未授權則回傳錯誤
+ */
+export const checkAuth = async (): Promise<CheckAuthResponse> => {
+  try {
+    console.log('發送請求: GET /api/auth/check');
+
+    // 使用 Next.js API 路由而不是直接呼叫後端 API
+    const res = await fetch('/api/auth/check', {
+      method: 'GET',
+      credentials: 'include', // 包含 Cookie
+    });
+
+    console.log('回應狀態:', res.status, res.statusText);
+
+    // 解析回應資料
+    const responseData = await res.json();
+    console.log('解析後的回應資料:', responseData);
+
+    // 檢查回應是否成功
+    if (responseData.Status === false) {
+      throw new Error(responseData.Message || '身份驗證失敗');
+    }
+
+    return responseData;
+  } catch (error) {
+    console.error('檢查使用者認證狀態失敗:', error);
+    throw error;
+  }
+};
+
+/**
+ * 使用電子郵件註冊新帳號
+ */
+export const registerWithEmail = async (
+  email: string,
+  name: string,
+  password: string,
+): Promise<RegisterResponse> => {
+  try {
+    console.log(`發送請求: POST ${apiConfig.baseUrl}/auth/register`);
+
+    const requestData = {
+      AccountEmail: email,
+      AccountName: name,
+      Password: password,
+    };
+
+    console.log('請求資料:', { ...requestData, Password: '***' });
+
+    const res = await fetch(`${apiConfig.baseUrl}/auth/register`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(requestData),
+    });
+
+    console.log('回應狀態:', res.status, res.statusText);
+
+    // 解析回應資料
+    const responseText = await res.text();
+    console.log('回應原始文本:', responseText);
+
+    let responseData;
+    try {
+      responseData = JSON.parse(responseText);
+      console.log('解析後的回應資料:', responseData);
+    } catch (e) {
+      console.error('解析 JSON 失敗:', e);
+      throw new Error(`回應不是有效的 JSON: ${responseText}`);
+    }
+
+    return responseData;
+  } catch (error) {
+    console.error('註冊失敗:', error);
+    return {
+      StatusCode: 500,
+      msg: error instanceof Error ? error.message : '註冊過程中發生未知錯誤',
+    };
+  }
+};
+
+/**
+ * 使用電子郵件與密碼登入帳號
+ */
+export const loginWithEmail = async (
+  email: string,
+  password: string,
+): Promise<LoginResponse> => {
+  try {
+    console.log(`發送請求: POST ${apiConfig.baseUrl}/auth/login`);
+
+    const requestData = {
+      AccountEmail: email,
+      Password: password,
+    };
+
+    console.log('請求資料:', { ...requestData, Password: '***' });
+
+    const res = await fetch(`${apiConfig.baseUrl}/auth/login`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(requestData),
+    });
+
+    console.log('回應狀態:', res.status, res.statusText);
+
+    // 解析回應資料
+    const responseText = await res.text();
+    console.log('回應原始文本:', responseText);
+
+    let responseData;
+    try {
+      responseData = JSON.parse(responseText);
+      console.log('解析後的回應資料:', responseData);
+    } catch (e) {
+      console.error('解析 JSON 失敗:', e);
+      throw new Error(`回應不是有效的 JSON: ${responseText}`);
+    }
+
+    // 如果登入成功，將 token 存儲到 localStorage
+    if (responseData.StatusCode === 200 && responseData.token) {
+      // 檢查是否在瀏覽器環境
+      if (typeof window !== 'undefined') {
+        updateAuthToken(responseData.token);
+
+        // 如果有用戶資料，也存儲到 localStorage
+        if (responseData.userData) {
+          localStorage.setItem(
+            'userData',
+            JSON.stringify(responseData.userData),
+          );
+        }
+      }
+    }
+
+    return responseData;
+  } catch (error) {
+    console.error('登入失敗:', error);
+    return {
+      StatusCode: 500,
+      msg: error instanceof Error ? error.message : '登入過程中發生未知錯誤',
+    };
+  }
+};
