+++ src/services/index.ts
@@ -1,5 +1,14 @@
-// 客戶端 API
-export * from './api';
+// 匯出 utils 工具
+export * from './utils/http';
+
+// 匯出認證服務模組
+export * from './auth';
+
+// 匯出食譜服務模組
+export * from './recipes';
+
+// 匯出用戶服務模組
+export * from './users';
 
 // 伺服器端 API - 明確匯出函數
 export {
