+++ src/services/recipes/index.ts
@@ -1,0 +1,1 @@
+export * from './recipe-service';
