+++ src/services/recipes/recipe-service.ts
@@ -1,107 +1,26 @@
-import { apiConfig, authConfig } from '@/config';
+import { apiConfig } from '@/config';
 import {
   Recipe,
-  RecipeFormData,
   ApiResponse,
-  GoogleAuthResponse,
+  RecipeFormData,
   RecipeCreateResponse,
   RecipeStep2Data,
   VideoUploadResponse,
   RecipeDraftResponse,
-  // RecipeDraftStep,
   UpdateStepsRequest,
   UpdateStepsResponse,
   SubmitDraftResponse,
   SubmitDraftDetail,
   SubmitDraftStep,
-  RegisterResponse,
-  LoginResponse,
-  CheckAuthResponse,
-  UserProfileResponse,
-  CurrentUserProfileResponse,
-  UpdateUserProfileResponse,
   AuthorRecipesResponse,
   DeleteMultipleResponse,
   TogglePublishResponse,
   UserFavoriteFollowResponse,
-  UserRecipesResponse,
-  FollowResponse,
   FavoriteRecipeResponse,
   RecipeRatingCommentResponse,
   RecipeTeachingResponse,
 } from '@/types/api';
-
-/**
- * 從 Cookie 或 localStorage 獲取 JWT Token
- */
-export const getAuthToken = (): string | null => {
-  // 開發環境下使用測試 token
-  if (process.env.NODE_ENV === 'development') {
-    console.log('開發環境：使用測試 token');
-    return 'eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJJZCI6MjksIkRpc3BsYXlJZCI6Ik0wMDAwMDIiLCJBY2NvdW50RW1haWwiOiJhMTIzQGdtYWlsLmNvbSIsIkFjY291bnROYW1lIjoiQWxpY2UiLCJSb2xlIjowLCJMb2dpblByb3ZpZGVyIjowLCJFeHAiOiIyMDI1LTA0LTI3VDEyOjM4OjA0LjIyNDg3OTlaIn0.MjTGyLcMjwBKq_BkySyPk2aIjfKmx_SzY8O3cLcRNYfY5ksh4oPbAXCTwYRTJTAANAzyGwC3F1siYfXh5FYl5g';
-  }
-
-  // 在伺服器端 document 不存在
-  if (typeof window === 'undefined') return null;
-
-  // 1. 嘗試從 localStorage 獲取 token (如果存在)
-  const localToken = localStorage.getItem('authToken');
-  if (localToken) {
-    return localToken;
-  }
-
-  // 2. 嘗試從 Cookie 獲取 Token (可能無法獲取 HttpOnly Cookie)
-  const cookies = document.cookie.split(';');
-  const authCookie = cookies
-    .map((cookie) => cookie.trim().split('='))
-    .find(([name]) => name === authConfig.tokenCookieName);
-
-  if (authCookie) {
-    return decodeURIComponent(authCookie[1]);
-  }
-
-  return null;
-};
-
-/**
- * 更新 Cookie 和 localStorage 中的 JWT Token
- */
-export const updateAuthToken = (token: string): void => {
-  if (typeof window === 'undefined') return;
-
-  // 1. 更新 Cookie (可能由服務器設置為 HttpOnly)
-  const expirationDate = new Date();
-  expirationDate.setDate(expirationDate.getDate() + authConfig.tokenExpiryDays);
-
-  document.cookie = `${authConfig.tokenCookieName}=${encodeURIComponent(
-    token,
-  )}; expires=${expirationDate.toUTCString()}; path=/; SameSite=Strict; Secure`;
-
-  // 2. 同時存儲到 localStorage 以便客戶端讀取
-  localStorage.setItem('authToken', token);
-};
-
-/**
- * 獲取 Google 登入 URL
- */
-export const fetchGoogleAuthUrl = async (): Promise<string> => {
-  try {
-    console.log(`發送請求: GET ${apiConfig.baseUrl}/auth/google/auth`);
-    const res = await fetch(`${apiConfig.baseUrl}/auth/google/auth`);
-    console.log('回應狀態:', res.status, res.statusText);
-
-    if (!res.ok) {
-      throw new Error(`HTTP error! status: ${res.status}`);
-    }
-
-    const data: GoogleAuthResponse = await res.json();
-    console.log('回應資料:', data);
-    return data.redirectUri;
-  } catch (error) {
-    console.error('獲取 Google 登入 URL 失敗:', error);
-    throw error;
-  }
-};
+import { getAuthToken } from '../utils/http';
 
 /**
  * 獲取首頁顯示的食譜列表
@@ -213,36 +132,6 @@ export const uploadRecipeBasic = async (
     return data;
   } catch (error) {
     console.error('上傳食譜失敗:', error);
-    throw error;
-  }
-};
-
-/**
- * 使用 Google 授權碼換取 token
- */
-export const exchangeGoogleCodeForToken = async (
-  code: string,
-): Promise<any> => {
-  try {
-    // 直接調用後端 API，而不是 Next.js API 路由
-    const res = await fetch(
-      `${apiConfig.baseUrl}/auth/google/callback?code=${encodeURIComponent(code)}`,
-      {
-        method: 'GET',
-      },
-    );
-
-    console.log('回應狀態:', res.status, res.statusText);
-
-    if (!res.ok) {
-      throw new Error(`HTTP error! status: ${res.status}`);
-    }
-
-    const data = await res.json();
-    console.log('回應資料:', data);
-    return data;
-  } catch (error) {
-    console.error('Google code 換取 token 失敗:', error);
     throw error;
   }
 };
@@ -569,293 +458,6 @@ export const submitRecipeDraft = async (
 };
 
 /**
- * 檢查使用者認證狀態並取得新的 Token 與使用者資料
- * @returns 包含新 Token 和使用者資料的回應，若未授權則回傳錯誤
- */
-export const checkAuth = async (): Promise<CheckAuthResponse> => {
-  try {
-    console.log('發送請求: GET /api/auth/check');
-
-    // 使用 Next.js API 路由而不是直接呼叫後端 API
-    const res = await fetch('/api/auth/check', {
-      method: 'GET',
-      credentials: 'include', // 包含 Cookie
-    });
-
-    console.log('回應狀態:', res.status, res.statusText);
-
-    // 解析回應資料
-    const responseData = await res.json();
-    console.log('解析後的回應資料:', responseData);
-
-    // 檢查回應是否成功
-    if (responseData.Status === false) {
-      throw new Error(responseData.Message || '身份驗證失敗');
-    }
-
-    return responseData;
-  } catch (error) {
-    console.error('檢查使用者認證狀態失敗:', error);
-    throw error;
-  }
-};
-
-/**
- * 使用電子郵件註冊新帳號
- */
-export const registerWithEmail = async (
-  email: string,
-  name: string,
-  password: string,
-): Promise<RegisterResponse> => {
-  try {
-    console.log(`發送請求: POST ${apiConfig.baseUrl}/auth/register`);
-
-    const requestData = {
-      AccountEmail: email,
-      AccountName: name,
-      Password: password,
-    };
-
-    console.log('請求資料:', { ...requestData, Password: '***' });
-
-    const res = await fetch(`${apiConfig.baseUrl}/auth/register`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(requestData),
-    });
-
-    console.log('回應狀態:', res.status, res.statusText);
-
-    // 解析回應資料
-    const responseText = await res.text();
-    console.log('回應原始文本:', responseText);
-
-    let responseData;
-    try {
-      responseData = JSON.parse(responseText);
-      console.log('解析後的回應資料:', responseData);
-    } catch (e) {
-      console.error('解析 JSON 失敗:', e);
-      throw new Error(`回應不是有效的 JSON: ${responseText}`);
-    }
-
-    return responseData;
-  } catch (error) {
-    console.error('註冊失敗:', error);
-    return {
-      StatusCode: 500,
-      msg: error instanceof Error ? error.message : '註冊過程中發生未知錯誤',
-    };
-  }
-};
-
-/**
- * 使用電子郵件與密碼登入帳號
- */
-export const loginWithEmail = async (
-  email: string,
-  password: string,
-): Promise<LoginResponse> => {
-  try {
-    console.log(`發送請求: POST ${apiConfig.baseUrl}/auth/login`);
-
-    const requestData = {
-      AccountEmail: email,
-      Password: password,
-    };
-
-    console.log('請求資料:', { ...requestData, Password: '***' });
-
-    const res = await fetch(`${apiConfig.baseUrl}/auth/login`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(requestData),
-    });
-
-    console.log('回應狀態:', res.status, res.statusText);
-
-    // 解析回應資料
-    const responseText = await res.text();
-    console.log('回應原始文本:', responseText);
-
-    let responseData;
-    try {
-      responseData = JSON.parse(responseText);
-      console.log('解析後的回應資料:', responseData);
-    } catch (e) {
-      console.error('解析 JSON 失敗:', e);
-      throw new Error(`回應不是有效的 JSON: ${responseText}`);
-    }
-
-    // 如果登入成功，將 token 存儲到 localStorage
-    if (responseData.StatusCode === 200 && responseData.token) {
-      // 檢查是否在瀏覽器環境
-      if (typeof window !== 'undefined') {
-        localStorage.setItem('authToken', responseData.token);
-
-        // 如果有用戶資料，也存儲到 localStorage
-        if (responseData.userData) {
-          localStorage.setItem(
-            'userData',
-            JSON.stringify(responseData.userData),
-          );
-        }
-      }
-    }
-
-    return responseData;
-  } catch (error) {
-    console.error('登入失敗:', error);
-    return {
-      StatusCode: 500,
-      msg: error instanceof Error ? error.message : '登入過程中發生未知錯誤',
-    };
-  }
-};
-
-/**
- * 取得使用者個人檔案資料
- */
-export const fetchUserProfile = async (
-  displayId: string,
-): Promise<UserProfileResponse> => {
-  try {
-    console.log(`發送請求: GET /api/user/${displayId}/profile`);
-
-    // 發送請求到 Next.js API 路由
-    const res = await fetch(`/api/user/${displayId}/profile`, {
-      method: 'GET',
-      credentials: 'include', // 包含 Cookie
-    });
-
-    console.log('回應狀態:', res.status, res.statusText);
-
-    // 解析回應資料
-    const responseText = await res.text();
-    console.log('回應原始文本:', responseText);
-
-    let responseData;
-    try {
-      responseData = JSON.parse(responseText);
-      console.log('回應資料:', responseData);
-    } catch (e) {
-      console.error('解析 JSON 失敗:', e);
-      throw new Error(`回應不是有效的 JSON: ${responseText}`);
-    }
-
-    return responseData;
-  } catch (error) {
-    console.error('獲取使用者資料失敗:', error);
-    throw error;
-  }
-};
-
-/**
- * 獲取當前登入使用者的個人資料
- * 只有登入的用戶可以使用此 API 查詢自身資料
- */
-export const fetchCurrentUserProfile =
-  async (): Promise<CurrentUserProfileResponse> => {
-    try {
-      console.log('發送請求: GET /api/user/profile');
-
-      // 使用 Next.js API 路由而不是直接呼叫後端 API
-      const res = await fetch('/api/user/profile', {
-        method: 'GET',
-        credentials: 'include', // 包含 Cookie
-      });
-
-      console.log('回應狀態:', res.status, res.statusText);
-
-      // 解析回應資料
-      const responseData = await res.json();
-      console.log('解析後的回應資料:', responseData);
-
-      // 處理錯誤狀態碼
-      if (responseData.StatusCode !== 200) {
-        throw new Error(responseData.msg || '獲取用戶資料失敗');
-      }
-
-      return responseData;
-    } catch (error) {
-      console.error('獲取當前用戶資料失敗:', error);
-      throw error;
-    }
-  };
-
-/**
- * 更新當前登入使用者的個人資料
- * @param data 要更新的用戶資料
- * @param profilePhoto 頭像照片檔案 (可選)
- */
-export const updateUserProfile = async (
-  data: {
-    accountName?: string;
-    description?: string;
-  },
-  profilePhoto?: File,
-): Promise<UpdateUserProfileResponse> => {
-  try {
-    console.log(`發送請求: PUT /api/user/profile`);
-    console.log('請求資料:', { ...data, profilePhoto: profilePhoto?.name });
-
-    // 創建 FormData 物件
-    const formData = new FormData();
-
-    // 添加資料
-    if (data.accountName) {
-      formData.append('accountName', data.accountName);
-    }
-
-    if (data.description) {
-      formData.append('description', data.description);
-    }
-
-    // 添加頭像照片，如果有
-    if (profilePhoto instanceof File) {
-      formData.append('profilePhoto', profilePhoto);
-    }
-
-    // 發送請求到 Next.js API 路由
-    const res = await fetch(`/api/user/profile`, {
-      method: 'PUT',
-      credentials: 'include', // 包含 Cookie
-      body: formData,
-    });
-
-    console.log('回應狀態:', res.status, res.statusText);
-
-    // 解析回應資料
-    const responseText = await res.text();
-    console.log('回應原始文本:', responseText);
-
-    let responseData;
-    try {
-      responseData = JSON.parse(responseText);
-      console.log('解析後的回應資料:', responseData);
-    } catch (e) {
-      console.error('解析 JSON 失敗:', e);
-      throw new Error(`回應不是有效的 JSON: ${responseText}`);
-    }
-
-    // 處理錯誤狀態碼
-    if (responseData.StatusCode !== 200) {
-      throw new Error(responseData.msg || '更新用戶資料失敗');
-    }
-
-    return responseData;
-  } catch (error) {
-    console.error('更新用戶資料失敗:', error);
-    throw error;
-  }
-};
-
-/**
  * 取得作者食譜列表
  * @param displayId 作者的顯示 ID (例如 U000001)
  * @param isPublished 是否只查詢已發布的食譜，false 為草稿
@@ -1061,135 +663,6 @@ export const fetchUserFavoriteFollow = async (
 };
 
 /**
- * 取得使用者的公開食譜列表
- */
-export const fetchUserRecipes = async (
-  displayId: string,
-  page: number = 1,
-): Promise<UserRecipesResponse> => {
-  try {
-    console.log(
-      `發送請求: GET ${apiConfig.baseUrl}/user/${displayId}/recipes?page=${page}`,
-    );
-
-    // 發送請求
-    const res = await fetch(
-      `${apiConfig.baseUrl}/user/${displayId}/recipes?page=${page}`,
-      {
-        method: 'GET',
-      },
-    );
-
-    console.log('回應狀態:', res.status, res.statusText);
-
-    // 處理 404 或其他錯誤狀態
-    if (!res.ok) {
-      if (res.status === 404) {
-        console.warn(`找不到使用者 ${displayId} 的食譜資料`);
-        // 返回空資料而非拋出錯誤
-        return {
-          statusCode: 404,
-          hasMore: false,
-          recipeCount: 0,
-          recipes: [],
-        };
-      }
-      throw new Error(`HTTP error! status: ${res.status}`);
-    }
-
-    // 讀取回應文字以便檢查和除錯
-    const responseText = await res.text();
-    if (!responseText.trim()) {
-      console.warn('API 回應內容為空');
-      return {
-        statusCode: 200,
-        hasMore: false,
-        recipeCount: 0,
-        recipes: [],
-      };
-    }
-
-    // 解析回應資料
-    try {
-      const data = JSON.parse(responseText);
-      console.log('回應資料:', data);
-
-      // 如果回應狀態不是成功
-      if (data.statusCode !== 200) {
-        console.warn(`API 回應狀態碼非 200: ${data.statusCode}`);
-        return {
-          statusCode: data.statusCode,
-          hasMore: false,
-          recipeCount: 0,
-          recipes: [],
-          message: data.msg || '獲取使用者食譜失敗',
-        };
-      }
-
-      return data;
-    } catch (parseError) {
-      console.error('解析 JSON 失敗:', parseError, '原始文本:', responseText);
-      throw new Error('解析回應資料失敗');
-    }
-  } catch (error) {
-    console.error('獲取使用者食譜失敗:', error);
-    throw error;
-  }
-};
-
-/**
- * 追蹤指定使用者
- */
-export const followUser = async (userId: number): Promise<FollowResponse> => {
-  try {
-    console.log(`發送請求: POST /api/users/${userId}/follow`);
-
-    // 發送請求到 Next.js API 路由
-    const res = await fetch(`/api/users/${userId}/follow`, {
-      method: 'POST',
-      credentials: 'include', // 包含 Cookie
-    });
-
-    console.log('回應狀態:', res.status, res.statusText);
-
-    // 解析回應資料
-    const data = await res.json();
-    console.log('回應資料:', data);
-
-    return data;
-  } catch (error) {
-    console.error('追蹤使用者失敗:', error);
-    throw error;
-  }
-};
-
-/**
- * 取消追蹤指定使用者
- */
-export const unfollowUser = async (userId: number): Promise<FollowResponse> => {
-  try {
-    console.log(`發送請求: DELETE /api/users/${userId}/follow`);
-
-    // 發送請求到 Next.js API 路由
-    const res = await fetch(`/api/users/${userId}/follow`, {
-      method: 'DELETE',
-      credentials: 'include', // 包含 Cookie
-    });
-
-    console.log('回應狀態:', res.status, res.statusText);
-
-    // 解析回應資料
-    const data = await res.json();
-    console.log('回應資料:', data);
-
-    return data;
-  } catch (error) {
-    console.error('取消追蹤使用者失敗:', error);
-    throw error;
-  }
-};
-
-/**
  * 收藏食譜
  */
 export const favoriteRecipe = async (
@@ -1370,6 +843,9 @@ export const submitRecipeRatingComment = async (
   }
 };
 
+/**
+ * 獲取食譜教學資訊
+ */
 export const fetchRecipeTeaching = async (
   recipeId: number,
 ): Promise<RecipeTeachingResponse> => {
